Replace an empty string translation with a space
In the English UI of Korp, if you do a trend diagram and some
part of the corpus is missing time data (like with Ylilauta), the UI
prints eg. "non_time_before99.81% of selected material lacks time data."

I now manually fixed this with a whitespace instead of an empty string
to reflect the contents if this commit, but this is not the only empty
string, so possibly others should be changed too.